% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Node.R
\name{all_ancestors}
\alias{all_ancestors}
\title{Find all ancestors of given node}
\usage{
all_ancestors(node_name, tmle_nodes)
}
\arguments{
\item{node_name}{the node to search for ancestors of}

\item{tmle_nodes}{the list of nodes and parents (i.e. the graph)}
}
\description{
Find all ancestors of given node
}
