% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Bound (Truncate) Likelihoods}
\usage{
bound(x, lower = 0.01, upper = 1 - lower)
}
\arguments{
\item{x}{the likelihood values to bound}

\item{lower}{the lower threshold}

\item{upper}{the upper threshold (or 1-lower by default)}
}
\description{
Bound (Truncate) Likelihoods
}
