% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF_likelihood.R
\docType{data}
\name{CF_likelihood}
\alias{CF_likelihood}
\title{Counterfactual likelihood}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CF_likelihood
}
\description{
Based on a sample_likelihood estimated from observed data.
Modifies likelihood factors according to intervention_list
}
\keyword{datasets}
