% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counterfactual.R
\docType{data}
\name{Counterfactual}
\alias{Counterfactual}
\title{Class for Defining Counterfactuals}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Counterfactual
}
\description{
Class for Defining Counterfactuals
}
\keyword{datasets}
