% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_base.R
\name{define_lf}
\alias{define_lf}
\title{Defining Likelihood Functionals}
\usage{
define_lf(LF_class, ...)
}
\arguments{
\item{LF_class}{the class of likelihood}

\item{...}{arguments that define the likelihood functional}
}
\description{
Defining Likelihood Functionals
}
