% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Node.R
\name{define_node}
\alias{define_node}
\title{Define a Node (set of variables) in an NPSEM}
\usage{
define_node(name, variables, parents = c(), variable_type = NULL)
}
\arguments{
\item{name}{A character, the name of node}

\item{variables}{A character vector of the variables that comprise the node}

\item{parents}{A character vector of the names of the parent nodes. If missing, node is assumed to have no parents.}

\item{variable_type}{A \code{sl3} \code{variable_type} object specifying the data type of this variable.
If missing, variable_type is guessed from the data}
}
\description{
Define a Node (set of variables) in an NPSEM
}
