% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_base.R
\name{define_param}
\alias{define_param}
\title{Define a Parameter}
\usage{
define_param(Param_class, ...)
}
\arguments{
\item{Param_class}{the class of the Parameter. Should inherit from \code{\link{Param_base}}}

\item{...}{arguments that define the parameter See the constructor for the specified \code{Parameter}.}
}
\description{
Define a Parameter
}
\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_stratified}},
\code{\link{Param_survival}},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
