% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_method}
\alias{delta_method}
\title{Get Inference for Differentiable Functions of Parameters}
\usage{
delta_method(delta_param, estimates, param_names = NULL)
}
\arguments{
\item{delta_param}{\code{list} of functions, defines the parameter}

\item{estimates}{\code{list}, TMLE estimates of parameter and ICs from \code{\link{tmle3_Fit}$estimates}}

\item{param_names}{\code{character} vector, the names of the parameters}
}
\value{
\code{list}, delta method estimates of parameter and ICS
}
\description{
Using the functional delta method, get estimates and inference for a smooth function of parameters
}
