% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_functions.R
\docType{data}
\name{delta_param_OR}
\alias{delta_param_OR}
\title{Odds Ratio odds(Y1)/odds(Y0)}
\format{
An object of class \code{list} of length 5.
}
\usage{
delta_param_OR
}
\description{
Odds Ratio odds(Y1)/odds(Y0)
}
\keyword{datasets}
