% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_functions.R
\docType{data}
\name{delta_param_PAF}
\alias{delta_param_PAF}
\title{PAF = 1 - (1/RR(EY/E0))}
\format{
An object of class \code{list} of length 5.
}
\usage{
delta_param_PAF
}
\description{
PAF = 1 - (1/RR(EY/E0))
}
\keyword{datasets}
