% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_functions.R
\docType{data}
\name{delta_param_RR}
\alias{delta_param_RR}
\title{Risk Ratio EY1/EY0}
\format{
An object of class \code{list} of length 5.
}
\usage{
delta_param_RR
}
\description{
Risk Ratio EY1/EY0
}
\keyword{datasets}
