% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_variable.R
\name{discretize_variable}
\alias{discretize_variable}
\title{Discretize Continuous Variable}
\usage{
discretize_variable(data, variable, num_cats, breakpoints = NULL)
}
\arguments{
\item{data}{\code{data.table}, containing the column to change}

\item{variable}{\code{character}, the name of the column to change}

\item{num_cats}{\code{integer}, the number of bins to generate}

\item{breakpoints}{\code{numeric vector}, the breakpoints to use. If NULL, these will be quantiles.}
}
\value{
the updated \code{data.table}, modified in place
}
\description{
Converts a \code{data.table} column from continuous to a discrete factor
}
