% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ED_from_estimates}
\alias{ED_from_estimates}
\title{Get Empirical Mean of EIFs from Estimates}
\usage{
ED_from_estimates(estimates)
}
\arguments{
\item{estimates}{a list of estimates objects}
}
\description{
Get Empirical Mean of EIFs from Estimates
}
