% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{f_contrast}
\alias{f_contrast}
\title{Linear Contrast Function}
\usage{
f_contrast(x)
}
\arguments{
\item{x}{numeric vector, containing two parameter estimates}
}
\value{
\code{x[[2]] - x[[1]]}
}
\description{
\eqn{X_2-X_1}
}
\details{
For use with \code{\link{delta_method}}
}
