% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{f_paf}
\alias{f_paf}
\alias{f_paf}
\title{RR Function}
\usage{
f_paf(x)

f_paf(x)
}
\arguments{
\item{x}{numeric vector, containing two parameter estimates}

\item{x}{numeric vector, containing two parameter estimates}
}
\value{
\code{x[[2]]/x[[1]]}

\code{D(x[[2]])/x[[1]]}
}
\description{
\eqn{\frac{X_2}{X_1}}

\eqn{\frac{X_2}{X_1}}
}
\details{
For use with \code{\link{delta_method}}

For use with \code{\link{delta_method}}
}
