% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{generate_loss_function_from_family}
\alias{generate_loss_function_from_family}
\title{Generate loss function loss from family object or string}
\usage{
generate_loss_function_from_family(family)
}
\arguments{
\item{family}{...}
}
\description{
Generate loss function loss from family object or string
}
