% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{generate_submodel_from_family}
\alias{generate_submodel_from_family}
\title{Generate Fluctuation Submodel from \code{family} object.}
\usage{
generate_submodel_from_family(family)
}
\arguments{
\item{family}{...}
}
\description{
Generate Fluctuation Submodel from \code{family} object.
}
