% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient.R
\docType{class}
\name{Gradient}
\alias{Gradient}
\title{Class representing a (possibly non-canonical) gradient for some parameter}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{likelihood}{A trained likelihood object from which to compute the gradient from.}

\item{projection_task_generator}{A function that takes a task, likelihood, target parameter, and target_node
and returns a task with the outcome being the evaluated gradient,
and covariates being whatever variables the conditional density of the target node likelihood factor depends on.
(E.g. covariates + outcome in the regression task)}
}
\value{
\code{Gradient} object
}
\description{
Class representing a (possibly non-canonical) gradient for some parameter
}
\keyword{data}
