% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_collapsed.R
\name{LF_binomial_collapsed}
\alias{LF_binomial_collapsed}
\title{Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LF_base} object
}
\description{
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
}
\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_base}},
\code{\link{LF_derived}},
\code{\link{LF_emp}},
\code{\link{LF_fit}},
\code{\link{LF_known}},
\code{\link{LF_static}},
\code{\link{LF_targeted}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}},
\code{\link{define_lf}()}
}
\concept{Likelihood objects}
\keyword{data}
