% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_emp.R
\name{LF_emp}
\alias{LF_emp}
\title{Likelihood Factor Estimated using Empirical Distribution}
\format{\code{\link{R6Class}} object.}
\value{
\code{LF_base} object
}
\description{
Uses the empirical probability distribution (puts mass \eqn{1/n} on each of the observations, or uses weights if specified) to estimate a marginal density.
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
Only compatible with marginal likelihoods (no parent nodes). Only compatible with densities (no conditional means).
The \code{type} argument will be ignored if specified.
}
\section{Constructor}{

  \code{define_lf(LF_emp, name, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{...}}{Not currently used.
    }
    }
}

\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_base}},
\code{\link{LF_derived}},
\code{\link{LF_fit}},
\code{\link{LF_known}},
\code{\link{LF_static}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}},
\code{\link{define_lf}()}
}
\concept{Likelihood objects}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:LF_base]{tmle3::LF_base}} -> \code{Lf_emp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LF_emp$new()}}
\item \href{#method-get_mean}{\code{LF_emp$get_mean()}}
\item \href{#method-get_density}{\code{LF_emp$get_density()}}
\item \href{#method-sample}{\code{LF_emp$sample()}}
\item \href{#method-clone}{\code{LF_emp$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="cf_values">}\href{../../tmle3/html/LF_base.html#method-cf_values}{\code{tmle3::LF_base$cf_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="delayed_train">}\href{../../tmle3/html/LF_base.html#method-delayed_train}{\code{tmle3::LF_base$delayed_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="get_likelihood">}\href{../../tmle3/html/LF_base.html#method-get_likelihood}{\code{tmle3::LF_base$get_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="print">}\href{../../tmle3/html/LF_base.html#method-print}{\code{tmle3::LF_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="train">}\href{../../tmle3/html/LF_base.html#method-train}{\code{tmle3::LF_base$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_emp$new(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\subsection{Method \code{get_mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_emp$get_mean(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_density"></a>}}
\subsection{Method \code{get_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_emp$get_density(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_emp$sample(tmle_task = NULL, n_samples = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_emp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
