% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_np.R
\docType{data}
\name{LF_np}
\alias{LF_np}
\title{nonparamteric likelihood (for baseline covariates only)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LF_np
}
\description{
nonparamteric likelihood (for baseline covariates only)
}
\keyword{datasets}
