% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_shift.R
\docType{data}
\name{LF_shift}
\alias{LF_shift}
\title{Shifted Likelihood Factor}
\format{\code{\link{R6Class}} object.}
\usage{
LF_shift
}
\value{
\code{LF_base} object
}
\description{
Shifts a likelihood factor according to a \code{shift_function}. 
In effect, \code{get_likelihood(tmle_task)} will instead the likelihood from the \code{original_lf},
but for shifted value \eqn{A'=}\code{shift_function}\eqn{(A,W)}
}
\section{Constructor}{

  \code{define_lf(LF_shift, name, type = "density", original_lf, shift_function, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$tmle_nodes}
    }
    \item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
    }
    \item{\code{original_lf}}{\code{\link{LF_base}} object, the likelihood factor to shift
    }
    \item{\code{shift_function}}{\code{function}, defines the shifted data
    }
    \item{\code{...}}{Not currently used.
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{original_lf}}{\code{\link{LF_base}} object, the likelihood factor to shift
    }
    \item{\code{shift_function}}{\code{function}, defines the shifted data
    }
    }
}

\references{
Díaz, Iván, and Mark J van der Laan. 2017. “Stochastic Treatment Regimes.” In Targeted Learning in Data Science: Causal Inference for Complex Longitudinal Studies, 167–80. Springer Science & Business Media.
Muñoz, Iván Díaz, and Mark J van der Laan. 2012. “Population Intervention Causal Effects Based on Stochastic Interventions.” Biometrics 68 (2). Wiley Online Library: 541–49.
}
\seealso{
Other Likelihood objects: \code{\link{CF_Likelihood}},
  \code{\link{LF_base}}, \code{\link{LF_fit}},
  \code{\link{LF_np}}, \code{\link{LF_static}},
  \code{\link{Likelihood}}, \code{\link{define_lf}}
}
\keyword{data}
