% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_static.R
\docType{data}
\name{LF_static}
\alias{LF_static}
\title{Defining Likelihood Functionals for Static Interventions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LF_static
}
\description{
Defining Likelihood Functionals for Static Interventions
}
\keyword{datasets}
