% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\docType{data}
\name{Likelihood}
\alias{Likelihood}
\title{Class for Likelihood Computations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Likelihood
}
\description{
Class for Likelihood Computations
}
\keyword{datasets}
