% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{logistic_loss}
\alias{logistic_loss}
\title{Logistic Loss Function}
\usage{
logistic_loss(estimate, observed)
}
\arguments{
\item{estimate}{...}

\item{observed}{...}
}
\description{
Logistic Loss Function
}
