% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loglik_loss}
\alias{loglik_loss}
\title{Log-likelihood Loss Function}
\usage{
loglik_loss(estimate, observed)
}
\arguments{
\item{estimate}{...}

\item{observed}{...}
}
\description{
Log-likelihood Loss Function
}
