% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loss_function_least_squares}
\alias{loss_function_least_squares}
\title{Least-squares loss for binary variables}
\usage{
loss_function_least_squares(
  estimate,
  observed,
  weights = NULL,
  likelihood = NULL
)
}
\arguments{
\item{estimate}{...}

\item{observed}{...}

\item{weights}{...}

\item{likelihood}{...}
}
\description{
Least-squares loss for binary variables
}
