% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loss_function_loglik_binomial}
\alias{loss_function_loglik_binomial}
\title{Log likelihood loss for binary variables}
\usage{
loss_function_loglik_binomial(
  estimate,
  observed,
  weights = NULL,
  likelihood = NULL
)
}
\arguments{
\item{estimate}{...}

\item{observed}{...}

\item{weights}{...}

\item{v}{...}
}
\description{
Log likelihood loss for binary variables
}
