% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loss_function_poisson}
\alias{loss_function_poisson}
\title{Poisson/log-linear loss for nonnegative variables}
\usage{
loss_function_poisson(
  estimate,
  observed,
  weights = NULL,
  likelihood = NULL,
  tmle_task = NULL,
  fold_number = NULL
)
}
\arguments{
\item{estimate}{...}

\item{observed}{...}

\item{weights}{...}

\item{likelihood}{...}
}
\description{
Poisson/log-linear loss for nonnegative variables
}
