% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loss_loglik}
\alias{loss_loglik}
\title{log likelihood loss}
\usage{
loss_loglik(estimate, observed)
}
\arguments{
\item{estimate}{prediction}

\item{observed}{observed outcome}
}
\description{
log likelihood loss
}
