% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loss_loglik_binomial}
\alias{loss_loglik_binomial}
\title{Log likelihood loss for outcomes between 0 and 1}
\usage{
loss_loglik_binomial(estimate, observed)
}
\arguments{
\item{estimate}{prediction}

\item{observed}{observed outcome}
}
\description{
Log likelihood loss for outcomes between 0 and 1
}
