% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{loss_poisson}
\alias{loss_poisson}
\title{Poisson/log-linear loss for nonnegative variables}
\usage{
loss_poisson(estimate, observed)
}
\arguments{
\item{estimate}{prediction}

\item{observed}{observed outcome}
}
\description{
Poisson/log-linear loss for nonnegative variables
}
