% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_glm_semiparametric.R
\docType{class}
\name{Lrnr_glm_semiparametric}
\alias{Lrnr_glm_semiparametric}
\title{Semiparametric Generalized Linear Models}
\format{
\code{\link{R6Class}} object.
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides fitting procedures for generalized linear models using
\code{\link[stats]{glm.fit}}.
}
\section{Parameters}{

\describe{
\item{\code{...}}{Parameters passed to \code{\link[stats]{glm}}.}
}
}

\concept{Learners}
\keyword{data}
