% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_tmle_submodel.R
\docType{data}
\name{Lrnr_tmle_update}
\alias{Lrnr_tmle_update}
\title{Fitting submodel}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Lrnr_tmle_update
}
\description{
Fitting submodel
}
\keyword{datasets}
