% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_define_blip.R
\name{make_blip_chain_function}
\alias{make_blip_chain_function}
\title{Learner Class for Blip Regressions}
\usage{
make_blip_chain_function(counterfactual_0, counterfactual_1, V)
}
\arguments{
\item{counterfactual_0}{A \code{Counterfactual} object defining one possible
intervention}

\item{counterfactual_1}{A \code{Counterfactual} object defining a second
possible intervention}

\item{V}{A character vector listing the nodes on which to regress the blip
binary blip for now}
}
\description{
Learner Class for Blip Regressions
}
