% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{make_submodel_spec}
\alias{make_submodel_spec}
\title{Main maker of submodel specs.}
\usage{
make_submodel_spec(
  name,
  family = NULL,
  submodel_function = NULL,
  loss_function = NULL
)
}
\arguments{
\item{name}{...}
}
\description{
Main maker of submodel specs.
}
