% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Node.R
\name{all_ancestors}
\alias{all_ancestors}
\alias{time_ordering}
\title{Helper functions for the NPSEM}
\usage{
all_ancestors(node_name, tmle_nodes)

time_ordering(tmle_nodes)
}
\arguments{
\item{node_name}{the node to search for ancestors of}

\item{tmle_nodes}{the NPSEM, defined by a list of \code{\link{tmle3_Node}} objects.}
}
\description{
\code{all_ancestors} returns a list of all_ancestors of the specified node.
\code{time_ordering} attempts to find a time_ordering for the variables.

Get time ordering of nodes
}
