% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_ATC.R
\name{Param_ATC}
\alias{Param_ATC}
\title{Additive Effect of Treatment Among the Treated}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Parameter definition for the Additive Effect of Treatment Among the Treated (ATT). Currently supports multiple static intervention nodes.
Does yet not support dynamic rule or stochastic interventions.
}
\section{Current Issues}{

\itemize{
\item clever covariates doesn't support updates; always uses initial (necessary for iterative TMLE, e.g. stochastic intervention)
\item doesn't integrate over possible counterfactuals (necessary for stochastic intervention)
\item clever covariate gets recalculated all the time (inefficient)
}
}

\section{Constructor}{

\code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
}
\item{\code{...}}{Not currently used.
}
\item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
}
\item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
}
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_stratified}},
\code{\link{Param_survival}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
