% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_ATE.R
\docType{data}
\name{Param_ATE}
\alias{Param_ATE}
\title{Class for Estimation of ATE Parameter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Param_ATE
}
\description{
Class for Estimation of ATE Parameter
}
\keyword{datasets}
