% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_base.R
\docType{data}
\name{Param_base}
\alias{Param_base}
\title{Base Class for Parameter Estimation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Param_base
}
\description{
Base Class for Parameter Estimation
}
\keyword{datasets}
