% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_coxph.R
\name{Param_coxph}
\alias{Param_coxph}
\title{Nonparametric inference for user-specified parametric working models for the conditional treatment effect.
The true conditional average treatment effect is projected onto a parametric working model using least-squares regression.
Unlike \code{Param_npCATT}, this function uses all observations to compute the projection.
This can be used to assess heterogeneity of the average treatment effect.
We note that \code{formula_coxph = ~ 1} gives an estimator of the nonparametric average treatment effect (ATE).}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Parameter definition for the Average Treatment Effect (ATE).
}
\section{Constructor}{

\code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link{Likelihood}} cocoxphesponding to the observed likelihood
}
\item{\code{formula_coxph}}{...
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
}
\item{\code{...}}{Not cucoxphently used.
}
\item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
}
\item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
}
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_npCATE}},
\code{\link{Param_npCATT}},
\code{\link{Param_npOR}},
\code{\link{Param_npRR}},
\code{\link{Param_npTSM}},
\code{\link{Param_spCATE}},
\code{\link{Param_spOR}},
\code{\link{Param_spRR}},
\code{\link{Param_stratified}},
\code{\link{Param_survival}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
