% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_CR.R
\name{Param_CR}
\alias{Param_CR}
\title{Average Treatment Effect}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Parameter definition for the Average Treatment Effect (ATE).
}
\section{Constructor}{

\code{define_param(Param_CR, observed_likelihood, intervention_list, marginalized = F, censoring_node = "processA", competing_risk_nodes, target_risk_node, target_times = NULL, node_time_ordering = NULL)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
}
\item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention.
}
\item{\code{marginalized}}{ Whether the initial estimation was and thus the targeting should be through the marginalized cause specific hazards
or the conditional (time-ordered) cause specific hazards.
}
\item{\code{censoring_node}} {The name of the censoring node
}
\item{\code{competing_risk_nodes}} {The name of the competing risk nodes (excluding the count process node of interest)}
\item{\code{target_risk_node}} {The counting process node of interest.}
\item{\code{target_times}} {The times to target}
\item{\code{node_time_ordering}} {The time ordering of the competing risk nodes. Only applicable if marginalized == F.}
\item{\code{...}}{Not currently used.}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_LATE}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base2}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_stratified}},
\code{\link{Param_survival_old}},
\code{\link{Param_survival}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
