% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_LATE.R
\name{Param_LATE}
\alias{Param_LATE}
\title{Longitudinal Average Treatment Effect
Parameter for ATE in (L0 A0 L1 A1 ... Y) data structure with any number of time dependent covariates and treatments.
Supports arbitrarily many time points.
Parameter definition for the Longitudinal Average Treatment Effect (LATE).}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Longitudinal Average Treatment Effect
Parameter for ATE in (L0 A0 L1 A1 ... Y) data structure with any number of time dependent covariates and treatments.
Supports arbitrarily many time points.
Parameter definition for the Longitudinal Average Treatment Effect (LATE).
}
\section{Constructor}{

\code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
}
\item{\code{...}}{Not currently used.
}
\item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
}
\item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
}
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_CR}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_middle_projection}},
\code{\link{Param_stratified}},
\code{\link{Param_survival_old}},
\code{\link{Param_survival}},
\code{\link{Param_thresh}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
