% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_mean.R
\name{Param_mean}
\alias{Param_mean}
\title{Mean of Outcome Node}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
Parameter for marginal mean of Y: \eqn{\Psi=E[Y]}. No TMLE update needed, but can be used in delta method calculations.
Useful for example, in calculating attributable risks.
}
\section{Constructor}{

  \code{define_param(Param_TSM, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this treatment
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention
    }
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_stratified}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_mean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_mean$new()}}
\item \href{#method-clever_covariates}{\code{Param_mean$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_mean$estimates()}}
\item \href{#method-clone}{\code{Param_mean$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_mean$new(observed_likelihood, ..., outcome_node = "Y")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_mean$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_mean$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_mean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
