% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_MSM.R
\name{Param_MSM}
\alias{Param_MSM}
\title{Stratified Parameter Estimates via MSM}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
Stratified Parameter Estimates via MSM

Stratified Parameter Estimates via MSM
}
\section{Current Issues}{

\itemize{
  \item clever covariates doesn't support updates; always uses initial (necessary for iterative TMLE, e.g. stochastic intervention)
  \item clever covariate gets recalculated all the time (inefficient)
}
}

\section{Constructor}{

  \code{define_param(Param_MSM, observed_likelihood, strata_variable, ...)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{msm}}{form of the MSM. Default is "A + V", consistent with the default of 
    \code{treatment_node} and \code{strata_name}.
    }
    \item{\code{weight}}{"Cond.Prob.", "Unif." or custom input function. 
    Note that custom function should support vector input. Default is "Cond.Prob.".
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{covariate_node}}{character, the name of the node that should be treated as the covariate
    }
    \item{\code{treatment_node}}{character, the name of the node that should be treated as the treatment
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this treatment
    }
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_stratified}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_MSM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_MSM$new()}}
\item \href{#method-clever_covariates}{\code{Param_MSM$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_MSM$estimates()}}
\item \href{#method-get_treatment_indicators}{\code{Param_MSM$get_treatment_indicators()}}
\item \href{#method-clone}{\code{Param_MSM$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_MSM$new(
  observed_likelihood,
  strata_variable,
  strata_name = "V",
  msm = "A + V",
  weight = "Cond.Prob.",
  weight_ub = 1/0.025,
  continuous_treatment = FALSE,
  treatment_values = NULL,
  n_samples = 30,
  ...,
  covariate_node = "W",
  treatment_node = "A",
  outcome_node = "Y"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_MSM$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_MSM$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_treatment_indicators"></a>}}
\subsection{Method \code{get_treatment_indicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_MSM$get_treatment_indicators(
  tmle_task,
  treatment_node = self$treatment_node
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_MSM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
