% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_npOR.R
\name{Param_npOR}
\alias{Param_npOR}
\title{Nonparametric inference for user-specified parametric working models for the conditional odds ratio between two binary variables
The true conditional odds ratio is projected onto a parametric working model using logistic-regression.
This can be used to assess heterogeneity of the odds ratio.
We note that \code{formula_logOR = ~ 1} gives an estimator of the nonparametric marginal odds ratio among the treated.
The parametric model is at the log-scale and therefore the coefficients returned code the linear predictor for the \code{log}-conditional odds ratio.
Parameter definition for the Average Treatment Effect (ATE).}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Nonparametric inference for user-specified parametric working models for the conditional odds ratio between two binary variables
The true conditional odds ratio is projected onto a parametric working model using logistic-regression.
This can be used to assess heterogeneity of the odds ratio.
We note that \code{formula_logOR = ~ 1} gives an estimator of the nonparametric marginal odds ratio among the treated.
The parametric model is at the log-scale and therefore the coefficients returned code the linear predictor for the \code{log}-conditional odds ratio.
Parameter definition for the Average Treatment Effect (ATE).
}
\section{Constructor}{

\code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
}
\item{\code{formula_logOR}}{...
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
}
\item{\code{...}}{Not currently used.
}
\item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
}
\item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
}
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_contCATE}},
\code{\link{Param_coxph}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_npCATE}},
\code{\link{Param_npCATT}},
\code{\link{Param_npRR}},
\code{\link{Param_npTSM}},
\code{\link{Param_spCATE}},
\code{\link{Param_spOR}},
\code{\link{Param_spRR}},
\code{\link{Param_stratified}},
\code{\link{Param_survival}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
