% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_spRR.R
\name{Param_spRR}
\alias{Param_spRR}
\title{Semiparametric estimation of the conditonal relative risk/treatment-effect for arbitrary partially-linear log-linear/link regression models.
Arbitrary user-specified parametric models for the conditional relative-risk are supported.
Assuming the semiparametric model to be true allows for some efficiency gain (when true) but may lead to less robust estimates due to misspecification.
The parametric model is at the log-scale and therefore the coefficients returned code the linear predictor for the \code{log}-relative-risk.}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Semiparametric estimation of the conditonal relative risk/treatment-effect for arbitrary partially-linear log-linear/link regression models.
Arbitrary user-specified parametric models for the conditional relative-risk are supported.
Assuming the semiparametric model to be true allows for some efficiency gain (when true) but may lead to less robust estimates due to misspecification.
The parametric model is at the log-scale and therefore the coefficients returned code the linear predictor for the \code{log}-relative-risk.
}
\section{Constructor}{

\code{define_param(Param_ATT, observed_likelihood, intervention_list, ..., outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
}
\item{\code{formula_logRR}}{...
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention.
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention.
}
\item{\code{...}}{Not currently used.
}
\item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood_treatment}}{the counterfactual likelihood for the treatment
}
\item{\code{cf_likelihood_control}}{the counterfactual likelihood for the control
}
\item{\code{intervention_list_treatment}}{A list of objects inheriting from \code{\link{LF_base}}, representing the treatment intervention
}
\item{\code{intervention_list_control}}{A list of objects inheriting from \code{\link{LF_base}}, representing the control intervention
}
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_npCATE}},
\code{\link{Param_npCATT}},
\code{\link{Param_npOR}},
\code{\link{Param_npRR}},
\code{\link{Param_npTSM}},
\code{\link{Param_spCATE}},
\code{\link{Param_spOR}},
\code{\link{Param_stratified}},
\code{\link{Param_survival}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
