% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim.R
\name{plot_vim}
\alias{plot_vim}
\title{Plot results of variable importance analysis}
\usage{
plot_vim(vim_results)
}
\arguments{
\item{vim_results}{Object produced by invoking \code{tmle3_vim}.}
}
\description{
Plot results of variable importance analysis
}
