% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middle_helpers.R, R/point_treatment_helpers.R
\name{middle_npsem}
\alias{middle_npsem}
\alias{middle_task}
\alias{middle_likelihood}
\alias{point_tx_npsem}
\alias{point_tx_task}
\alias{point_tx_likelihood}
\title{Helper Functions for Longitudinal Mediation}
\usage{
middle_npsem(node_list, variable_types = NULL)

middle_task(data, node_list, variable_types = NULL, ...)

middle_likelihood(tmle_task, learner_list)

point_tx_npsem(node_list, variable_types = NULL)

point_tx_task(data, node_list, variable_types = NULL, ...)

point_tx_likelihood(tmle_task, learner_list)
}
\arguments{
\item{node_list}{a list of character vectors, listing the variables that comprise each node}

\item{variable_types}{a list of variable types, one for each node. If missing, variable types will be guessed}

\item{data}{a \code{data.frame}, or \code{data.table} containing data for use in estimation}

\item{...}{extra arguments.}

\item{tmle_task}{a \code{\link{tmle3_Task}} as constructed via \code{point_tx_task}}

\item{learner_list}{a list of sl3 learners, one for A and one for Y to be used for likelihood estimation}
}
\description{
Handles the ARZL(Y) structure

Handles the common W (covariates), A (treatment/intervention), Y (outcome) data structure
}
