% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_missing.R
\name{process_missing}
\alias{process_missing}
\title{Preprocess Data to Handle Missing Variables}
\usage{
process_missing(data, node_list, max_p_missing = 0.5)
}
\arguments{
\item{data, }{\code{data.table}, containing the missing variables}

\item{node_list, }{\code{list}, what nodes serve what purpose}

\item{max_p_missing, }{\code{numeric}, what proportion of missing is tolerable? Beyond that, the variable will be dropped from the analysis}
}
\value{
\code{list} containing the following elements:
 \itemize{
   \item \code{data}, the updated dataset
   \item \code{node_list}, the updated list of nodes
   \item \code{n_dropped}, the number of observations dropped
   \item \code{dropped_cols}, the variables dropped due to excessive missingness

 }
}
\description{
Process data to account for missingness in preparation for TMLE
}
\details{
Because IPCW-TMLE is not yet implemented, currently all rows with
missing outcomes (Y) or treatments (A) will be dropped from the analysis.

Then covariates will be processed as follows:
\enumerate{
 \item any covariate with more than \code{max_p_missing} missingness will be dropped
 \item indicators of missingness will be generated
 \item missing values will be median-imputed
}
}
