% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_scores.R
\name{density_formula}
\alias{density_formula}
\alias{get_propensity_scores}
\alias{propensity_score_plot}
\alias{propensity_score_table}
\title{Get and Plot Propensity Scores}
\usage{
density_formula(tmle_task, node = "A")

get_propensity_scores(likelihood, tmle_task, node = "A")

propensity_score_plot(likelihood, tmle_task, node = "A")

propensity_score_table(likelihood, tmle_task, node = "A")
}
\arguments{
\item{tmle_task}{a tmle_task data object}

\item{node}{a character specifing which node to use}

\item{likelihood}{a fitted likelihood object}
}
\description{
Get and Plot Propensity Scores
}
