% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_density}
\alias{submodel_density}
\title{Density Submodel Fluctuation
Standard (1+ eps D)P submodel through densities}
\usage{
submodel_density(eps, X, offset)
}
\arguments{
\item{eps}{...}

\item{X}{...}

\item{offset}{...}
}
\description{
Density Submodel Fluctuation
Standard (1+ eps D)P submodel through densities
}
