% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_exp}
\alias{submodel_exp}
\title{Log-linear (Poisson) Submodel Fluctuation}
\usage{
submodel_exp(eps, X, offset)
}
\arguments{
\item{eps}{...}

\item{X}{...}

\item{offset}{...}
}
\description{
Log-linear (Poisson) Submodel Fluctuation
}
