% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_exponential}
\alias{submodel_exponential}
\title{Exponential Submodel Fluctuation (e.g. continuous hazard/partial likelihood)
Standard exp(log(offset) + eps H) submodel}
\usage{
submodel_exponential(eps, offset, X)
}
\arguments{
\item{eps}{...}

\item{offset}{...}

\item{X}{...}
}
\description{
Exponential Submodel Fluctuation (e.g. continuous hazard/partial likelihood)
Standard exp(log(offset) + eps H) submodel
}
