% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_linear}
\alias{submodel_linear}
\title{Linear (gaussian) Submodel Fluctuation}
\usage{
submodel_linear(eps, offset, X, observed)
}
\arguments{
\item{eps}{...}

\item{offset}{...}

\item{X}{...}
}
\description{
Linear (gaussian) Submodel Fluctuation
}
