% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_logistic}
\alias{submodel_logistic}
\title{Logistic Submodel Fluctuation}
\usage{
submodel_logistic(eps, offset, X)
}
\arguments{
\item{eps}{...}

\item{offset}{...}

\item{X}{...}
}
\description{
Logistic Submodel Fluctuation
}
