% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_logistic_switch}
\alias{submodel_logistic_switch}
\title{Logistic Submodel Fluctuation for likelihood (not conditional means)}
\usage{
submodel_logistic_switch(eps, offset, X, observed)
}
\arguments{
\item{eps}{...}

\item{offset}{...}

\item{X}{...}
}
\description{
Logistic Submodel Fluctuation for likelihood (not conditional means)
}
