% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodel_spec}
\alias{submodel_spec}
\title{Function that returns the necessary updating information/functions for
a given optimization type in the targetting step.
Currently supports targetting through logistic submodels or (1 + eps D) P likelihood submodels}
\usage{
submodel_spec(optim_type = c("logistic", "EIC"))
}
\arguments{
\item{optim_type}{A targeting method for the optimization.}
}
\description{
Function that returns the necessary updating information/functions for
a given optimization type in the targetting step.
Currently supports targetting through logistic submodels or (1 + eps D) P likelihood submodels
}
