% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\docType{data}
\name{submodel_spec_logistic_switch}
\alias{submodel_spec_logistic_switch}
\title{Submodel for binary outcomes where "initial" is a likelihood and not a conditional mean (e.g. for Param_ATC and Param_ATT for updating node \code{A}).}
\format{
An object of class \code{list} of length 4.
}
\usage{
submodel_spec_logistic_switch
}
\description{
Submodel for binary outcomes where "initial" is a likelihood and not a conditional mean (e.g. for Param_ATC and Param_ATT for updating node \code{A}).
}
\keyword{datasets}
