% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{summary_from_estimate}
\alias{summary_from_estimate}
\title{Summarize Estimate}
\usage{
summary_from_estimate(estimate)
}
\arguments{
\item{estimate}{\code{list}, TMLE estimates of parameter and ICs from \code{\link{tmle3_Fit}$estimates}}
}
\value{
\code{data.table} summarizing results
}
\description{
Generates a \code{data.table} summarizing results with inference
}
