% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_helpers.R
\name{survival_tx_npsem}
\alias{survival_tx_npsem}
\alias{survival_tx_task}
\alias{survival_tx_likelihood}
\title{Helper Functions for Survival Analysis}
\usage{
survival_tx_npsem(node_list, variable_types = NULL)

survival_tx_task(data, node_list, make_npsem, variable_types = NULL, ...)

survival_tx_likelihood(tmle_task, learner_list)
}
\arguments{
\item{node_list}{a list of character vectors, listing the variables that comprise each node}

\item{variable_types}{a list of variable types, one for each node. If missing, variable types will be guessed}

\item{data}{a \code{data.frame}, or \code{data.table} containing data for use in estimation}

\item{...}{extra arguments.}

\item{tmle_task}{a \code{\link{tmle3_Task}} as constructed via \code{survival_tx_task}}

\item{learner_list}{a list of sl3 learners, one for A and one for Y to be used for likelihood estimation}
}
\description{
Handles the W (covariates), A (treatment/intervention), T_tilde (time-to-event),
Delta (censoring indicator), t_max (the maximum time to estimate) survival data structure
}
