% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Node.R
\name{time_ordering}
\alias{time_ordering}
\title{Get time ordering of nodes}
\usage{
time_ordering(tmle_nodes)
}
\arguments{
\item{tmle_nodes}{the list of nodes and parents (i.e. the graph)}
}
\description{
Get time ordering of nodes
}
