% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3.R
\name{tmle3}
\alias{tmle3}
\title{TMLE from a tmle3_Spec object}
\usage{
tmle3(tmle_spec, data, node_list, learner_list = NULL)
}
\arguments{
\item{tmle_spec}{\code{\link{tmle3_Spec}}, defines the TMLE}

\item{data}{\code{data.frame}, the raw data}

\item{node_list}{\code{list}, defines which variables are which nodes}

\item{learner_list}{\code{list}, defines which learners are used to fit which likelihood factors}
}
\value{
A \code{\link{tmle3_Fit}} object
}
\description{
Using a tmle3_Spec object, fit a TMLE
}
