% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3.R
\docType{data}
\name{tmle3_Fit}
\alias{tmle3_Fit}
\title{Fit Targeted Maximum Likelihood Estimator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Fit
}
\arguments{
\item{likelihood}{...}

\item{task}{...}

\item{param}{...}

\item{lrnr_submodel}{...}
}
\description{
Fit Targeted Maximum Likelihood Estimator
}
\keyword{datasets}
