% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Node.R
\docType{data}
\name{tmle3_Node}
\alias{tmle3_Node}
\title{Class for Nonparametric Structural Equation Models}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Node
}
\description{
Class for Nonparametric Structural Equation Models
}
\keyword{datasets}
