% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec.R
\name{tmle3_Spec}
\alias{tmle3_Spec}
\title{Defines a TML Estimator (except for the data)}
\description{
Current limitations: pretty much tailored to \code{Param_TSM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tmle3_Spec$new()}}
\item \href{#method-make_tmle_task}{\code{tmle3_Spec$make_tmle_task()}}
\item \href{#method-make_initial_likelihood}{\code{tmle3_Spec$make_initial_likelihood()}}
\item \href{#method-make_updater}{\code{tmle3_Spec$make_updater()}}
\item \href{#method-make_targeted_likelihood}{\code{tmle3_Spec$make_targeted_likelihood()}}
\item \href{#method-make_params}{\code{tmle3_Spec$make_params()}}
\item \href{#method-clone}{\code{tmle3_Spec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$new(likelihood_override = NULL, variable_types = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_tmle_task"></a>}}
\subsection{Method \code{make_tmle_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$make_tmle_task(data, node_list, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_initial_likelihood"></a>}}
\subsection{Method \code{make_initial_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$make_initial_likelihood(tmle_task, learner_list = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_updater"></a>}}
\subsection{Method \code{make_updater()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$make_updater(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_targeted_likelihood"></a>}}
\subsection{Method \code{make_targeted_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$make_targeted_likelihood(likelihood, updater)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_params"></a>}}
\subsection{Method \code{make_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$make_params(tmle_task, targeted_likelihood)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
