% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_ADSM.R
\name{tmle3_Spec_ADSM}
\alias{tmle3_Spec_ADSM}
\title{Defines a TML Estimator for the Mean Outcome under a Candidate Adaptive Design}
\description{
Defines a TML Estimator for the Mean Outcome under a Candidate Adaptive Design
}
