% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_marginal_ADATE.R
\name{tmle3_Spec_marginal_ADATE}
\alias{tmle3_Spec_marginal_ADATE}
\title{Defines a TML Estimator for Marginal Average Treatment Effect under Adaptive Design}
\description{
Defines a TML Estimator for Marginal Average Treatment Effect under Adaptive Design
}
