% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_stratified.R
\docType{data}
\name{tmle3_Spec_stratified}
\alias{tmle3_Spec_stratified}
\title{Defines a TML Estimator (except for the data)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_stratified
}
\description{
Defines a TML Estimator (except for the data)
}
\keyword{datasets}
