% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Task.R
\docType{data}
\name{tmle3_Task}
\alias{tmle3_Task}
\title{Class for Storing and Computing TMLEs as Tasks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Task
}
\description{
Class for Storing and Computing TMLEs as Tasks
}
\keyword{datasets}
