% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_universal_Update.R
\docType{data}
\name{tmle3_universal_Update}
\alias{tmle3_universal_Update}
\title{Defines an update (submodel+loss function)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_universal_Update
}
\description{
Current Limitations:
only updating one node
loss function and submodel are hard-coded (need to accept arguments for these)
no support for one-step (recursive TMLE)
}
\keyword{datasets}
