% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim.R
\name{tmle3_vim}
\alias{tmle3_vim}
\title{Compute Variable Importance Measures (VIM) with any given parameter}
\usage{
tmle3_vim(
  tmle_spec,
  data,
  node_list,
  learner_list = NULL,
  adjust_for_other_A = TRUE
)
}
\arguments{
\item{tmle_spec}{\code{\link{tmle3_Spec}}, defines the TMLE}

\item{data}{\code{data.frame}, the raw data}

\item{node_list}{\code{list}, defines which variables are which nodes}

\item{learner_list}{\code{list}, defines which learners are used to fit which
likelihood factors}

\item{adjust_for_other_A}{Whether or not to adjust for other specified
intervention nodes.}
}
\description{
Compute Variable Importance Measures (VIM) with any given parameter
}
