% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_ATC.R
\name{tmle_ATC}
\alias{tmle_ATC}
\title{All Treatment Specific Means}
\usage{
tmle_ATC(treatment_level, control_level)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
