% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_ATE.R
\name{tmle_ATE}
\alias{tmle_ATE}
\title{All Treatment Specific Means}
\usage{
tmle_ATE(treatment_level, control_level)
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
