% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_marginal_ADATE.R
\name{tmle_marginal_ADATE}
\alias{tmle_marginal_ADATE}
\title{Mean Outcome under a Candidate Adaptive Design}
\usage{
tmle_marginal_ADATE(treatment_level, control_level, g_treat)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}

\item{g_treat}{the actual probability of A that corresponds to treatment}
}
\description{
O = (W, A, Y)
W = Covariates
A = Treatment (binary or categorical)
Y = Outcome (binary or bounded continuous)
}
