% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_marginal_ADSM.R
\name{tmle_marginal_ADSM}
\alias{tmle_marginal_ADSM}
\title{Mean Outcome under a Candidate Adaptive Design (Marginalized Over Covaraite Distribution)}
\usage{
tmle_marginal_ADSM(treatment_level, control_level, g_treat, g_adapt)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}

\item{g_treat}{the actual probability of A that corresponds to treatment}

\item{g_adapt}{the probability of A that corresponds to treatment under a candidate adaptive design}
}
\description{
O = (W, A, Y)
W = Covariates
A = Treatment (binary or categorical)
Y = Outcome (binary or bounded continuous)
}
