% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_middle.R
\name{tmle_middle}
\alias{tmle_middle}
\title{Longitudinal Mediation Targets}
\usage{
tmle_middle(treatment_level, control_level)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}
}
\description{
O=(L0, A1, R1, Z1, L1, (Y1), ..., An, Rn, Zn, Ln, Yn)
L0=Baseline covariates
A: Treatment (binary or categorical)
Z: Mediators
R: (time-varying) Covariates before mediator
L: (time-varying) Covariates after mediator
Y=Outcome (binary or bounded continuous)
}
