% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_MSM.R
\name{tmle_MSM}
\alias{tmle_MSM}
\title{Make MSM version of Stratified TML estimator class}
\usage{
tmle_MSM(mass = NULL)
}
\arguments{
\item{mass}{h(A, V)}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
