% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_OR.R
\name{tmle_OR}
\alias{tmle_OR}
\title{Odds Ratio}
\usage{
tmle_OR(baseline_level, contrast_level)
}
\arguments{
\item{baseline_level}{The baseline risk group.}

\item{contrast_level}{The contrast risk group.}
}
\description{
O = (W, A, Y)
W = Covariates
A = Treatment (binary or categorical)
Y = Outcome (binary or bounded continuous)
}
