% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_PAR.R
\name{tmle_PAR}
\alias{tmle_PAR}
\title{PAR and PAF}
\usage{
tmle_PAR(baseline_level)
}
\arguments{
\item{baseline_level, }{the baseline risk group}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
