% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_RR.R
\name{tmle_RR}
\alias{tmle_RR}
\title{Risk ratio}
\usage{
tmle_RR(baseline_level, contrast_level)
}
\arguments{
\item{baseline_level, }{the baseline risk group}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
