% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_survival.R
\name{tmle_survival}
\alias{tmle_survival}
\title{Treatment Specific Survival}
\usage{
tmle_survival(treatment_level, control_level, target_times = NULL, ...)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}

\item{target_times}{the time points to be targeted at during the TMLE adjustment}

\item{...}{others args passed to spec}
}
\description{
See the associated handbook chapter
}
