% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_Threshold.R
\name{tmle_Threshold}
\alias{tmle_Threshold}
\title{All Treatment Specific Means}
\usage{
tmle_Threshold(threshold_values, num_thresholds, cdf_bins, ...)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
