% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec.R
\name{tmle_tsm_all}
\alias{tmle_tsm_all}
\title{All Treatment Specific Means}
\usage{
tmle_tsm_all()
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
