% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Bound (Truncate) Likelihoods}
\usage{
bound(x, bounds)
}
\arguments{
\item{x}{the likelihood values to bound}

\item{bounds}{Either a length two vector of c(lower,upper) or a lower bound, where the upper is then 1 - lower}
}
\description{
Bound (Truncate) Likelihoods
}
