% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counterfactual.R
\name{define_cf}
\alias{define_cf}
\title{Defining Counterfactuals}
\usage{
define_cf(intervention_list, name = NULL)
}
\arguments{
\item{intervention_list}{...}

\item{name}{...}
}
\description{
Defining Counterfactuals
}
