% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_base.R
\name{define_lf}
\alias{define_lf}
\title{Define a Likelihood Factor}
\usage{
define_lf(LF_class, ...)
}
\arguments{
\item{LF_class}{the class of likelihood factor. Should inherit from \code{\link{LF_base}}}

\item{...}{arguments that define the likelihood factor. See the constructor for the specified \code{LF_class}.}
}
\description{
Define a Likelihood Factor
}
\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_base}},
\code{\link{LF_derived}},
\code{\link{LF_emp}},
\code{\link{LF_fit}},
\code{\link{LF_known}},
\code{\link{LF_static}},
\code{\link{LF_targeted}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}}
}
\concept{Likelihood objects}
