% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_method}
\alias{delta_method}
\title{Get Inference for Differentiable Functions of Parameters}
\usage{
delta_method(estimates, f, df)
}
\arguments{
\item{estimates}{\code{list}, TMLE estimates of parameter and ICs from \code{\link{tmle3_Fit}$estimates}}

\item{f}{\code{function}, the function of parameter values}

\item{df}{\code{function}, the derivative of the function wrt the parameter values}
}
\value{
\code{list}, delta method estimates of parameter and ICS
}
\description{
Using the functional delta method, get estimates and inference for a smooth function of parameters
}
