% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_functions.R
\docType{data}
\name{delta_param_PAR}
\alias{delta_param_PAR}
\title{PAR = Linear Contrast EY-EY0}
\format{
An object of class \code{list} of length 4.
}
\usage{
delta_param_PAR
}
\description{
PAR = Linear Contrast EY-EY0
}
\keyword{datasets}
