% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_base.R
\docType{data}
\name{LF_base}
\alias{LF_base}
\title{Base Class for Defining Likelihoods}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LF_base
}
\description{
Base Class for Defining Likelihoods
}
\keyword{datasets}
