% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_base.R
\name{LF_base}
\alias{LF_base}
\title{Base Class for Defining Likelihood Factors}
\format{\code{\link{R6Class}} object.}
\value{
\code{LF_base} object
}
\description{
A Likelihood factor models a conditional density function.
The conditioning set is defined as all parent nodes (defined in \code{\link{tmle3_Task}}). In the case of a continuous
outcome variable, where a full density isn't needed, this can also model a conditional mean. This is the base class, which
is intended to be abstract. See below for a list of possible likelihood factor classes.
}
\section{Constructor}{

  \code{define_lf(LF_base, name, ..., type = "density")}
                     
  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
    }
    }
}

\section{Methods}{


\describe{
\item{\code{get_density(tmle_task)}}{
  Get conditional density values for for the observations in \code{tmle_task}.

  \itemize{
    \item{\code{tmle_task}: \code{\link{tmle3_Task}} to get likelihood values for
    }
  }
  }
\item{\code{get_mean(tmle_task)}}{
  Get conditional mean values for for the observations in \code{tmle_task}.
  \itemize{
    \item{\code{tmle_task}: \code{\link{tmle3_Task}} to get likelihood values for
    }
  }
  }
}
}

\section{Fields}{

\describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
    }
    \item{\code{variable_type}}{\code{\link[sl3]{variable_type}} object, specifying the data type of the outcome variable. Only available after Likelihood training.
    }
    \item{\code{values}}{Possible values of the outcome variable, retrivied from the \code{variable_type} object. Only available after Likelihood training.
    }
}
}

\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_derived}},
\code{\link{LF_emp}},
\code{\link{LF_fit}},
\code{\link{LF_known}},
\code{\link{LF_static}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}},
\code{\link{define_lf}()}
}
\concept{Likelihood objects}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LF_base$new()}}
\item \href{#method-delayed_train}{\code{LF_base$delayed_train()}}
\item \href{#method-train}{\code{LF_base$train()}}
\item \href{#method-get_density}{\code{LF_base$get_density()}}
\item \href{#method-get_mean}{\code{LF_base$get_mean()}}
\item \href{#method-get_likelihood}{\code{LF_base$get_likelihood()}}
\item \href{#method-sample}{\code{LF_base$sample()}}
\item \href{#method-cf_values}{\code{LF_base$cf_values()}}
\item \href{#method-print}{\code{LF_base$print()}}
\item \href{#method-clone}{\code{LF_base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$new(name, bound = NULL, ..., type = "density", cache = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delayed_train"></a>}}
\subsection{Method \code{delayed_train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$delayed_train(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$train(tmle_task, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_density"></a>}}
\subsection{Method \code{get_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$get_density(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\subsection{Method \code{get_mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$get_mean(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihood"></a>}}
\subsection{Method \code{get_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$get_likelihood(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$sample(tmle_task, n_samples = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cf_values"></a>}}
\subsection{Method \code{cf_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$cf_values(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
