% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_emp.R
\name{LF_emp}
\alias{LF_emp}
\title{Likelihood Factor Estimated using Empirical Distribution}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LF_base} object
}
\description{
Uses the empirical probability distribution (puts mass \eqn{1/n} on each of the observations, or uses weights if specified) to estimate a marginal density.
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
Only compatible with marginal likelihoods (no parent nodes). Only compatible with densities (no conditional means).
The \code{type} argument will be ignored if specified.
}
\section{Constructor}{

\code{define_lf(LF_emp, name, ...)}

\describe{
\item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
}
\item{\code{...}}{Not currently used.
}
}
}

\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_base}},
\code{\link{LF_derived}},
\code{\link{LF_fit}},
\code{\link{LF_known}},
\code{\link{LF_static}},
\code{\link{LF_targeted}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}},
\code{\link{define_lf}()}
}
\concept{Likelihood objects}
\keyword{data}
