% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_fit.R
\docType{data}
\name{LF_fit}
\alias{LF_fit}
\title{Fitting Likelihood Functionals}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LF_fit
}
\description{
Fitting Likelihood Functionals
}
\keyword{datasets}
