% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_known.R
\name{LF_known}
\alias{LF_known}
\title{Known True Likelihood Factor}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LF_base} object
}
\description{
Incorporate existing knowledge about the likelihood
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
}
\section{Constructor}{

\code{define_lf(LF_fit, name, mean_fun, density_fun, ..., type = "density")}

\describe{
\item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
}
\item{\code{mean_fun}}{A function that takes a sl3 regression task and returns true conditional means
}
\item{\code{density_fun}}{A function that takes a sl3 regression task and returns true conditional densities
}
\item{\code{...}}{Not currently used.
}
\item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
}
}
}

\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_base}},
\code{\link{LF_derived}},
\code{\link{LF_emp}},
\code{\link{LF_fit}},
\code{\link{LF_static}},
\code{\link{LF_targeted}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}},
\code{\link{define_lf}()}
}
\concept{Likelihood objects}
\keyword{data}
