% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_static.R
\name{LF_static}
\alias{LF_static}
\title{Static Likelihood Factor}
\format{\code{\link{R6Class}} object.}
\value{
\code{LF_base} object
}
\description{
Likelihood factor for a variable that only has one value with probability 1. This is used for static interventions.
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
}
\section{Constructor}{

  \code{define_lf(LF_static, name, type, value, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
    }
    \item{\code{value}}{the static value
    }
    \item{\code{...}}{Not currently used.
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{value}}{the static value.}
    }
}

\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood}},
\code{\link{LF_base}},
\code{\link{LF_derived}},
\code{\link{LF_emp}},
\code{\link{LF_fit}},
\code{\link{LF_known}},
\code{\link{Likelihood}},
\code{\link{Targeted_Likelihood}},
\code{\link{define_lf}()}
}
\concept{Likelihood objects}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:LF_base]{tmle3::LF_base}} -> \code{LF_static}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LF_static$new()}}
\item \href{#method-get_mean}{\code{LF_static$get_mean()}}
\item \href{#method-get_density}{\code{LF_static$get_density()}}
\item \href{#method-cf_values}{\code{LF_static$cf_values()}}
\item \href{#method-clone}{\code{LF_static$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="delayed_train">}\href{../../tmle3/html/LF_base.html#method-delayed_train}{\code{tmle3::LF_base$delayed_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="get_likelihood">}\href{../../tmle3/html/LF_base.html#method-get_likelihood}{\code{tmle3::LF_base$get_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="print">}\href{../../tmle3/html/LF_base.html#method-print}{\code{tmle3::LF_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="train">}\href{../../tmle3/html/LF_base.html#method-train}{\code{tmle3::LF_base$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_static$new(name, type = "density", value, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\subsection{Method \code{get_mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_static$get_mean(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_density"></a>}}
\subsection{Method \code{get_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_static$get_density(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cf_values"></a>}}
\subsection{Method \code{cf_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_static$cf_values(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_static$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
