% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_cache.R
\docType{class}
\name{Likelihood_cache}
\alias{Likelihood_cache}
\title{Cache Likelihood values, update those values}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Likelihood_cache
}
\description{
Cache Likelihood values, update those values
}
\keyword{datasets}
