% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_glm_semiparametric.R
\docType{class}
\name{Lrnr_glm_semiparametric}
\alias{Lrnr_glm_semiparametric}
\title{Semiparametric Generalized Linear Models}
\format{
\code{\link{R6Class}} object.
}
\value{
Learner object with methods for training and prediction. See
\code{\link{Lrnr_base}} for documentation on learners.
}
\description{
This learner provides fitting procedures for semiparametric generalized linear models using a user-given baseline learner and
\code{\link[stats]{glm.fit}}.
}
\section{Parameters}{

\describe{
\item{\code{formula_sp}}{ A \code{formula} object specifying the parametric component of the semiparametric model.}
\item{\code{lrnr_baseline}}{A baseline learner for estimation of the nonparametric component.}
\item{\code{interaction_variable}}{A interaction variable to multiply with the design matrix generated by \code{formula_sp}. If NULL then the interaction variable is treated as the value 1.
In many applications, this represents a binary treatment variable \code{A}.}
\item{\code{family}}{A family object whose link function specifies the type of semiparametric model (e.g. partially-linear least-squares (\code{gaussian}), partially-linear logistic regression (\code{binomial}), partially-linear relative-risk regression (\code{poisson}) }
\item{\code{append_interaction_matrix}}{Whether to \code{lrnr_baseline} should be fit on \code{cbind(task$X,V)} where \code{V} is the design matrix obtained from \code{formula_sp}.
Note, if \code{append_interaction_matrix = TRUE}, the resulting estimator will be projected onto the semiparametric model using \code{glm.fit}.
If this is FALSE and \code{interaction_variable} is binary then the semiparametric model is learned by stratifying on \code{interaction_variable}.
Specifically, if FALSE, \code{lrnr_baseline} is used to estimate \verb{E[Y|A=0,W]} by subsetting to only observations with \code{A} = 0.
In the binary case, setting \code{append_interaction_matrix = TRUE} allows one to pool the learning across treatment arms and allows additive models to perform well.  }
\item{\code{return_matrix_predictions}}{Only used if \code{interaction_variable} is binary. Whether to return a matrix output with three columns being \verb{E[Y|A=0,W], E[Y|A=1,W], E[Y|A,W]}.}

}
}

\concept{Learners}
\keyword{data}
