% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Node.R
\name{all_ancestors}
\alias{all_ancestors}
\alias{time_ordering}
\title{Helper functions for the NPSEM}
\usage{
all_ancestors(node_name, npsem)

time_ordering(npsem)
}
\arguments{
\item{node_name}{the node to search for ancestors of}

\item{npsem}{the NPSEM, defined by a list of \code{\link{tmle3_Node}} objects.}
}
\description{
\code{all_ancestors} returns a list of all_ancestors of the specified node.
\code{time_ordering} attempts to find a time_ordering for the variables.
}
