% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_base.R
\name{Param_base}
\alias{Param_base}
\title{Base Class for Defining Parameters}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
A parameter is a function of the likelihood. Once given a \code{\link{Likelihood}} object, a parameter will a value.
These objects also contain information about the efficient influence function (EIF) of a parameter, as well as its clever covariate(s).
}
\section{Constructor}{

  \code{define_param(Param_base, observed_likelihood, ..., outcome_node)}
                     
  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Methods}{


\describe{
\item{\code{clever_covariates(tmle_task = NULL)}}{
  Get the clever covariates for an TMLE update step.

  \itemize{
    \item{\code{tmle_task}: \code{\link{tmle3_Task}} to get clever covariate values for. 
    If NULL, the \code{tmle_task} used to train the \code{observed likelihood will be used}
    }
  }
  }
\item{\code{estimates(tmle_task = NULL)}}{
  Get the parameter estimates and influence curve values.

  \itemize{
    \item{\code{tmle_task}: \code{\link{tmle3_Task}} to get clever covariate values for. 
    If NULL, the \code{tmle_task} used to train the \code{observed likelihood will be used}
    }
  }
  }
}
}

\section{Fields}{

\describe{
    \item{\code{observed_likelihood}}{the observed likelihood
    }
    \item{\code{outcome_node}}{character, the name of the outcome node
    }
}
}

\seealso{
Other Parameters: 
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_stratified}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_base$new()}}
\item \href{#method-clever_covariates}{\code{Param_base$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_base$estimates()}}
\item \href{#method-print}{\code{Param_base$print()}}
\item \href{#method-clone}{\code{Param_base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$new(observed_likelihood, ..., outcome_node = "Y")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
