% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_CR.R
\name{Param_CR}
\alias{Param_CR}
\title{Competing Risks cause-specific cumulative incidence (no ties)
Supports multiple intervention nodes, and multiple competing risks + censoring.
Parameter definition for the CR cumulative incidence.}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Competing Risks cause-specific cumulative incidence (no ties)
Supports multiple intervention nodes, and multiple competing risks + censoring.
Parameter definition for the CR cumulative incidence.
}
\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_LATE}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_delta}},
\code{\link{Param_mean}},
\code{\link{Param_middle_projection}},
\code{\link{Param_stratified}},
\code{\link{Param_survival_old}},
\code{\link{Param_survival}},
\code{\link{Param_thresh}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
