% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_delta.R
\name{Param_delta}
\alias{Param_delta}
\title{Delta Method Parameters}
\description{
These parameters are smooth functionals of one or more other params
They are not fit directly with tmle, but are estimated using the delta method
todo: better docs
They do not return have clever covariates
}
\seealso{
Other Parameters: 
\code{\link{Param_ATC}},
\code{\link{Param_ATE}},
\code{\link{Param_ATT}},
\code{\link{Param_MSM}},
\code{\link{Param_TSM}},
\code{\link{Param_base}},
\code{\link{Param_mean}},
\code{\link{Param_stratified}},
\code{\link{Param_survival}},
\code{\link{define_param}()},
\code{\link{tmle3_Fit}}
}
\concept{Parameters}
\keyword{data}
