% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_TSM.R
\docType{data}
\name{Param_TSM}
\alias{Param_TSM}
\title{Class for Estimation of TSM Parameter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Param_TSM
}
\description{
Class for Estimation of TSM Parameter
}
\keyword{datasets}
