% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_from_estimates}
\alias{summary_from_estimates}
\title{Summarize Estimates}
\usage{
summary_from_estimates(
  task,
  estimates,
  param_types = NULL,
  param_names = NULL,
  init_psi = NULL,
  simultaneous_ci = FALSE
)
}
\arguments{
\item{task}{\code{tmle3_Task} containing the observed data of interest; the
same as that passed to ..}

\item{estimates}{\code{list}, TMLE estimates of parameter and ICs from
\code{\link{tmle3_Fit}$estimates}}

\item{param_types}{the types of the parameters being estimated}

\item{param_names}{the names of the parameters being estimated}

\item{init_psi}{the names of the parameters being estimated}

\item{simultaneous_ci}{if TRUE, calculate simulatenous confidence intervals}
}
\value{
\code{data.table} summarizing results
}
\description{
Generates a \code{data.table} summarizing results with inference
}
