% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec.R
\docType{data}
\name{tmle3_Spec}
\alias{tmle3_Spec}
\title{Defines a tmle (minus the data)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec
}
\description{
Current limitations:
pretty much tailored to Param_TSM
}
\keyword{datasets}
