% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_MSM.R
\docType{data}
\name{tmle3_Spec_MSM}
\alias{tmle3_Spec_MSM}
\title{Defines a Stratified TML Estimator with MSM (except for the data)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_MSM
}
\description{
Defines a Stratified TML Estimator with MSM (except for the data)
}
\keyword{datasets}
