% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_OR.R
\name{tmle3_Spec_OR}
\alias{tmle3_Spec_OR}
\title{Defines a TML Estimator for the Odds Ratio}
\description{
Current limitations:
pretty much tailored to Param_TSM
see TODOs for places generalization can be added
}
