% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_PAR.R
\name{tmle3_Spec_PAR}
\alias{tmle3_Spec_PAR}
\title{Defines a tmle (minus the data)}
\description{
Current limitations:
pretty much tailored to Param_TSM
see TODOs for places generalization can be added
}
