% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_TSM_all.R
\name{tmle3_Spec_TSM_all}
\alias{tmle3_Spec_TSM_all}
\title{Defines a TML Estimator (except for the data)}
\description{
Current limitations: pretty much tailored to \code{Param_TSM}
See TODOs for places generalization can be added
}
