% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Update.R
\name{tmle3_Update}
\alias{tmle3_Update}
\title{Defines an update procedure (submodel+loss function)}
\description{
Current Limitations:
loss function and submodel are hard-coded (need to accept arguments for these)
}
\section{Constructor}{

  \code{define_param(maxit, cvtmle, one_dimensional, constrain_step, delta_epsilon, verbose)}

  \describe{
    \item{\code{maxit}}{The maximum number of update iterations
    }
    \item{\code{cvtmle}}{If \code{TRUE}, use CV-likelihood values when
       calculating updates.
    }
    \item{\code{one_dimensional}}{If \code{TRUE}, collapse clever covariates
       into a one-dimensional clever covariate scaled by the mean of their
       EIFs.
    }
    \item{\code{constrain_step}}{If \code{TRUE}, step size is at most
       \code{delta_epsilon} (it can be smaller if a smaller step decreases
       the loss more).
    }
    \item{\code{delta_epsilon}}{The maximum step size allowed if
       \code{constrain_step} is \code{TRUE}.
    }
    \item{\code{convergence_type}}{The convergence criterion to use: (1)
       \code{"scaled_var"} corresponds to sqrt(Var(D)/n)/logn (the default)
       while (2) \code{"sample_size"} corresponds to 1/n.
    }
    \item{\code{fluctuation_type}}{Whether to include the auxiliary covariate
       for the fluctuation model as a covariate or to treat it as a weight.
       Note that the option \code{"weighted"} is incompatible with a
       multi-epsilon submodel (\code{one_dimensional = FALSE}).
    }
    \item{\code{verbose}}{If \code{TRUE}, diagnostic output is generated
       about the updating procedure.
    }
    }
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tmle3_Update$new()}}
\item \href{#method-collapse_covariates}{\code{tmle3_Update$collapse_covariates()}}
\item \href{#method-update_step}{\code{tmle3_Update$update_step()}}
\item \href{#method-generate_submodel_data}{\code{tmle3_Update$generate_submodel_data()}}
\item \href{#method-fit_submodel}{\code{tmle3_Update$fit_submodel()}}
\item \href{#method-submodel}{\code{tmle3_Update$submodel()}}
\item \href{#method-loss_function}{\code{tmle3_Update$loss_function()}}
\item \href{#method-apply_submodel}{\code{tmle3_Update$apply_submodel()}}
\item \href{#method-apply_update}{\code{tmle3_Update$apply_update()}}
\item \href{#method-check_convergence}{\code{tmle3_Update$check_convergence()}}
\item \href{#method-update}{\code{tmle3_Update$update()}}
\item \href{#method-register_param}{\code{tmle3_Update$register_param()}}
\item \href{#method-clone}{\code{tmle3_Update$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$new(
  maxit = 100,
  cvtmle = TRUE,
  one_dimensional = FALSE,
  constrain_step = FALSE,
  delta_epsilon = 1e-04,
  convergence_type = c("scaled_var", "sample_size"),
  fluctuation_type = c("standard", "weighted"),
  optim_delta_epsilon = TRUE,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collapse_covariates"></a>}}
\subsection{Method \code{collapse_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$collapse_covariates(estimates, clever_covariates)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_step"></a>}}
\subsection{Method \code{update_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$update_step(likelihood, tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_submodel_data"></a>}}
\subsection{Method \code{generate_submodel_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$generate_submodel_data(
  likelihood,
  tmle_task,
  fold_number = "full",
  update_node = "Y"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_submodel"></a>}}
\subsection{Method \code{fit_submodel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$fit_submodel(submodel_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submodel"></a>}}
\subsection{Method \code{submodel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$submodel(epsilon, initial, H)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loss_function"></a>}}
\subsection{Method \code{loss_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$loss_function(estimate, observed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_submodel"></a>}}
\subsection{Method \code{apply_submodel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$apply_submodel(submodel_data, epsilon)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_update"></a>}}
\subsection{Method \code{apply_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$apply_update(
  tmle_task,
  likelihood,
  fold_number,
  new_epsilon,
  update_node
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_convergence"></a>}}
\subsection{Method \code{check_convergence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$check_convergence(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$update(likelihood, tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register_param"></a>}}
\subsection{Method \code{register_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$register_param(new_params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Update$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
