% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_TSM_all.R
\name{tmle_TSM_all}
\alias{tmle_TSM_all}
\title{All Treatment Specific Means}
\usage{
tmle_TSM_all(...)
}
\arguments{
\item{...}{Extra arguments passed to the constructor of the superclass.}
}
\description{
O=(W,A,Y)
W=Covariates
A=Treatment (binary or categorical)
Y=Outcome (binary or bounded continuous)
}
